Add ExpandedNavigationBar and NavBarScrollViewWithOffset
