+++ Backpack-SwiftUI/NavigationBar/Classes/NavBarScrollViewWithOffset.swift
@@ -1,0 +1,57 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2018 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import SwiftUI
+
+struct NavBarScrollViewWithOffset<Content: View>: View {
+    let style: BPKNavigationBarStyle
+    let onScroll: (_ offset: CGPoint) -> Void
+    @ViewBuilder let content: () -> Content
+    
+    var body: some View {
+        ScrollView(showsIndicators: false) {
+            ZStack(alignment: .top) {
+                GeometryReader { geo in
+                    Color.clear
+                        .preference(
+                            key: ScrollOffsetPreferenceKey.self,
+                            value: geo.frame(in: .named("scrollView")).origin
+                        )
+                }
+                .frame(height: 0)
+                content()
+                    .offset(y: style.verticalOffset)
+            }
+        }
+        .coordinateSpace(name: "scrollView")
+        .onPreferenceChange(
+            ScrollOffsetPreferenceKey.self,
+            perform: onScroll
+        )
+        .if(style != .transparent, transform: { view in
+            view.safeAreaInset(edge: .bottom) {
+                Color.clear.frame(height: style.verticalOffset - style.largeTitlePadding)
+            }
+        })
+    }
+}
+
+private struct ScrollOffsetPreferenceKey: PreferenceKey {
+    static var defaultValue: CGPoint = .zero
+    static func reduce(value: inout CGPoint, nextValue: () -> CGPoint) {}
+}
