+++ Backpack-SwiftUI/Tests/NavBar/BPKNavigationViewTests.swift
@@ -51,7 +51,6 @@ class BPKNavigationViewTests: XCTestCase {
                 .clipped()
             Text("Screen content")
         }
-        .ignoresSafeArea(edges: .top)
     }
     
     func test_default() {
